Relocate notary client to notary folder.
