+++ .github/scripts/build-server.sh
@@ -6,7 +6,7 @@ environment=$1
 
 aws s3 sync .git s3://tlsn-deploy/$environment/.git --delete
 
-cd notary-server
+cd notary/server
 cargo build --release
 aws s3 cp target/release/notary-server s3://tlsn-deploy/$environment/
