+++ .github/workflows/cd-server.yml
@@ -49,8 +49,8 @@ jobs:
         uses: lewagon/wait-on-check-action@v1.3.1
         with:
           ref: ${{ github.ref }}
-          # Have to be specify '(notary-server)', as we are using matrix for build_and_test job in ci.yml, else it will fail, more details [here](https://github.com/lewagon/wait-on-check-action#check-name)
-          check-name: 'Build and test (notary-server)'
+          # Have to be specify '(notary)', as we are using matrix for build_and_test job in ci.yml, else it will fail, more details [here](https://github.com/lewagon/wait-on-check-action#check-name)
+          check-name: 'Build and test (notary)'
           repo-token: ${{ secrets.GITHUB_TOKEN }}
           # How frequent (in seconds) this job will call GitHub API to check the status of the job specified at 'check-name'
           wait-interval: 60
