--- notary-server/Cargo.toml
@@ -1,52 +1,0 @@
-[package]
-name = "notary-server"
-version = "0.1.0-alpha.5"
-edition = "2021"
-
-[dependencies]
-async-trait = "0.1.67"
-async-tungstenite = { version = "0.25", features = ["tokio-native-tls"] }
-axum = { version = "0.7", features = ["ws"] }
-axum-core = "0.4"
-axum-macros = "0.4"
-base64 = "0.21.0"
-chrono = "0.4.31"
-csv = "1.3.0"
-eyre = "0.6.8"
-futures = "0.3"
-futures-util = "0.3.28"
-http = "1.1"
-http-body-util = "0.1"
-hyper = { version = "1.1", features = ["client", "http1", "server"] }
-hyper-util = {version = "0.1", features = ["full"]}
-notify = { version = "6.1.1", default-features = false, features = ["macos_kqueue"] }
-p256 = "0.13"
-rustls = { version = "0.21" }
-rustls-pemfile = { version = "1.0.2" }
-serde = { version = "1.0.147", features = ["derive"] }
-serde_json = "1.0"
-serde_yaml = "0.9.21"
-sha1 = "0.10"
-structopt = "0.3.26"
-thiserror = "1"
-tlsn-verifier = { path = "../tlsn/tlsn-verifier", features = ["tracing"] }
-tokio = { version = "1", features = ["full"] }
-tokio-rustls = { version = "0.24.1" }
-tokio-util = { version = "0.7", features = ["compat"] }
-tower = { version = "0.4.12", features = ["make"] }
-tower-http = { version = "0.5", features = ["cors"] }
-tower-service = "0.3.2"
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-uuid = { version = "1.4.1", features = ["v4", "fast-rng"] }
-ws_stream_tungstenite = { version = "0.13.0", features = ["tokio_io"] }
-
-[dev-dependencies]
-# specify vendored feature to use statically linked copy of OpenSSL
-hyper-tls = { version = "0.6.0", features = ["vendored"] }
-rstest = "0.18"
-tlsn-notary-client = { path = "../tlsn/tlsn-notary-client" }
-tlsn-prover = { path = "../tlsn/tlsn-prover", features = ["tracing"] }
-tls-server-fixture = { path = "../components/tls/tls-server-fixture" }
-tlsn-tls-core = { path = "../components/tls/tls-core" }
-tokio-native-tls = { version = "0.3.1", features = ["vendored"] }
