+++ notary/client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tlsn-notary-client"
+name = "notary-client"
 version = "0.1.0-alpha.5"
 edition = "2021"
 
@@ -12,17 +12,20 @@ tracing = [
 ]
 
 [dependencies]
-notary-server = { path = "../../notary-server" }
-tlsn-tls-client.workspace = true
-tlsn-common.workspace = true
-tlsn-prover.workspace = true
+tlsn-tls-client = { path = "../../components/tls/tls-client" }
+tlsn-common = { path = "../../tlsn/tlsn-common" }
 
-derive_builder.workspace = true
 eyre.workspace = true
 http-body-util.workspace = true
 hyper = { workspace = true, features = ["client", "http1"] }
 hyper-util = { workspace = true, features = ["full"] }
+notary-server.workspace = true
+rustls.workspace = true
+rustls-pemfile.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
 thiserror.workspace = true
+tlsn-prover.workspace = true
 tokio = { workspace = true, features = [
   "rt",
   "rt-multi-thread",
@@ -31,12 +34,9 @@ tokio = { workspace = true, features = [
   "io-std",
   "fs",
 ] }
+tokio-rustls.workspace = true
 tokio-util.workspace = true
 tracing = { workspace = true, optional = true }
-webpki-roots.workspace = true
 
-rustls = "0.21"
-rustls-pemfile = "1.0.2"
-serde = { version = "1.0.147", features = ["derive"] }
-serde_json = "1.0"
-tokio-rustls = "0.24.1"
+derive_builder = "0.12"
+webpki-roots = "0.26"
