+++ notary/client/src/error.rs
