+++ notary/client/src/lib.rs
