+++ notary/integration-tests/tests/notary.rs
@@ -8,11 +8,11 @@ use hyper_util::{
     client::legacy::{connect::HttpConnector, Builder},
     rt::{TokioExecutor, TokioIo},
 };
+use notary_client::client::NotaryClient;
 use rstest::rstest;
 use std::{string::String, time::Duration};
 use tls_core::anchors::RootCertStore as TlsClientRootCertStore;
 use tls_server_fixture::{bind_test_server_hyper, CA_CERT_DER, SERVER_DOMAIN};
-use tlsn_notary_client::client::NotaryClient;
 use tlsn_prover::tls::{state::Setup, Prover, ProverConfig};
 use tokio_util::compat::{FuturesAsyncReadCompatExt, TokioAsyncReadCompatExt};
 use tracing::debug;
@@ -24,7 +24,7 @@ use notary_server::{
     NotarySigningKeyProperties, ServerProperties, TLSProperties,
 };
 
-const NOTARY_CA_CERT_BYTES: &[u8] = include_bytes!("../fixture/tls/rootCA.crt");
+const NOTARY_CA_CERT_BYTES: &[u8] = include_bytes!("../../server/fixture/tls/rootCA.crt");
 const MAX_SENT: usize = 1 << 13;
 const MAX_RECV: usize = 1 << 13;
 const API_KEY: &str = "test_api_key_0";
@@ -42,12 +42,12 @@ fn get_server_config(port: u16, tls_enabled: bool, auth_enabled: bool) -> Notary
         },
         tls: TLSProperties {
             enabled: tls_enabled,
-            private_key_pem_path: "./fixture/tls/notary.key".to_string(),
-            certificate_pem_path: "./fixture/tls/notary.crt".to_string(),
+            private_key_pem_path: "../server/fixture/tls/notary.key".to_string(),
+            certificate_pem_path: "../server/fixture/tls/notary.crt".to_string(),
         },
         notary_key: NotarySigningKeyProperties {
-            private_key_pem_path: "./fixture/notary/notary.key".to_string(),
-            public_key_pem_path: "./fixture/notary/notary.pub".to_string(),
+            private_key_pem_path: "../server/fixture/notary/notary.key".to_string(),
+            public_key_pem_path: "../server/fixture/notary/notary.pub".to_string(),
         },
         logging: LoggingProperties {
             level: "DEBUG".to_string(),
@@ -55,7 +55,7 @@ fn get_server_config(port: u16, tls_enabled: bool, auth_enabled: bool) -> Notary
         },
         authorization: AuthorizationProperties {
             enabled: auth_enabled,
-            whitelist_csv_path: "./fixture/auth/whitelist.csv".to_string(),
+            whitelist_csv_path: "../server/fixture/auth/whitelist.csv".to_string(),
         },
     }
 }
