+++ notary/server/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "notary-server"
+version = "0.1.0-alpha.5"
+edition = "2021"
+
+[dependencies]
+async-tungstenite = { workspace = true, features = ["tokio-native-tls"] }
+eyre.workspace = true
+http.workspace = true
+http-body-util.workspace = true
+hyper = { workspace = true, features = ["client", "http1", "server"] }
+hyper-util = { workspace = true, features = ["full"] }
+rustls.workspace = true
+rustls-pemfile.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace =true
+thiserror.workspace = true
+tokio = { workspace = true, features = ["full"] }
+tokio-rustls.workspace = true
+tokio-util = { workspace = true, features = ["compat"] }
+tracing.workspace = true
+tracing-subscriber = { workspace = true, features = ["env-filter"] }
+uuid = { workspace = true, features = ["v4", "fast-rng"] }
+ws_stream_tungstenite = { workspace = true, features = ["tokio_io"] }
+
+async-trait = "0.1.67"
+axum = { version = "0.7", features = ["ws"] }
+axum-core = "0.4"
+axum-macros = "0.4"
+base64 = "0.21.0"
+chrono = "0.4.31"
+csv = "1.3.0"
+futures = "0.3"
+futures-util = "0.3.28"
+notify = { version = "6.1.1", default-features = false, features = ["macos_kqueue"] }
+p256 = "0.13"
+serde_yaml = "0.9.21"
+sha1 = "0.10"
+structopt = "0.3.26"
+tlsn-verifier = { path = "../../tlsn/tlsn-verifier", features = ["tracing"] }
+tower = { version = "0.4.12", features = ["make"] }
+tower-http = { version = "0.5", features = ["cors"] }
+tower-service = "0.3.2"
