+++ notary/server/src/domain.rs
