+++ notary/server/src/domain/auth.rs
