+++ notary/server/src/domain/cli.rs
