+++ notary/server/src/domain/notary.rs
