+++ notary/server/src/error.rs
