+++ notary/server/src/lib.rs
