+++ notary/server/src/main.rs
