+++ notary/server/src/middleware.rs
