+++ notary/server/src/server.rs
