+++ notary/server/src/util.rs
