+++ tlsn/examples/discord/discord_dm.rs
@@ -1,6 +1,6 @@
 // This example shows how to notarize Discord DMs.
 //
-// The example uses the notary server implemented in ../../../notary-server
+// The example uses the notary server implemented in ../../../notary/server
 
 use http_body_util::{BodyExt, Empty};
 use hyper::{body::Bytes, Request, StatusCode};
@@ -15,7 +15,7 @@ use tracing::debug;
 // Setting of the application server
 const SERVER_DOMAIN: &str = "discord.com";
 
-// Setting of the notary server — make sure these are the same with those in ../../../notary-server
+// Setting of the notary server — make sure these are the same with those in ../../../notary/server
 const NOTARY_HOST: &str = "127.0.0.1";
 const NOTARY_PORT: u16 = 7047;
