+++ tlsn/examples/simple/simple_verifier.rs
@@ -69,7 +69,7 @@ fn main() {
 /// Returns a Notary pubkey trusted by this Verifier
 fn notary_pubkey() -> p256::PublicKey {
     let pem_file = str::from_utf8(include_bytes!(
-        "../../../notary-server/fixture/notary/notary.pub"
+        "../../../notary/server/fixture/notary/notary.pub"
     ))
     .unwrap();
     p256::PublicKey::from_public_key_pem(pem_file).unwrap()
