+++ tlsn/examples/twitter/README.md
@@ -29,13 +29,13 @@ Next, open the **Developer Tools**, go to the **Network** tab, and refresh the p
 ## Start the notary server
 At the root level of this repository, run 
 ```sh
-cd notary-server
+cd notary/server
 cargo run --release
 ```
 
 The notary server will now be running in the background waiting for connections.
 
-For more information on how to configure the notary server, please refer to [this](../../../notary-server/README.md#running-the-server).
+For more information on how to configure the notary server, please refer to [this](../../../notary/server/README.md#running-the-server).
 
 ## Notarize
